fixes CKW-4403
The cause of this bug was due to that fact that element id for the
group_selectedItems and sendee_selectedItems were hardcoded and when
salesforce regenerated the identifiers, the javascript responsible for
pull in checked items from the counter part elements failed.

This fix harmonizes this by using the $Component construct to generate
part of the identifier of the elements thereby giving a predicatable
identifier.